[FEATURE] Опция отключения файл-сервера + автокопирование binaries (тоже опция). (#236)
* [FIX] Убраны лишние классы из libLauncher.

* [FIX] Получение RAM.

* [FIX] Ещё фиксы такового с RAM.

* [REFACTOR] Удалён лишний код.

* [FEATURE] Теперь прикрутить sentry можно просто добавив библиотеку в libraries и дописав 3 строчки в runtime.

* [FIX] Ошибка в паттернах ServerWrapper, порождает java.util.IllegalFormatConversionException

* [FEATURE] Чистка конфига.

* [FIX] Убрал лишние варнинги.

* [FIX] Мои ошибки.

+++ Launcher/src/main/java/ru/gravit/launcher/client/FunctionalBridge.java
@@ -56,9 +56,9 @@ public class FunctionalBridge {
     }
 
     @LauncherAPI
-    public static long getTotalMemory() {
-    	if (cachedMemorySize > 0) return cachedMemorySize;
-    	return cachedMemorySize = hwidProvider.getTotalMemory() >> 20;
+    public static int getTotalMemory() {
+    	if (cachedMemorySize > 0) return (int)cachedMemorySize;
+    	return (int)(cachedMemorySize = hwidProvider.getTotalMemory() >> 20);
     }
 
     @LauncherAPI
@@ -67,7 +67,7 @@ public class FunctionalBridge {
     }
 
     @LauncherAPI
-    public static long getJVMTotalMemory() {
+    public static int getJVMTotalMemory() {
         if (getClientJVMBits() == 32) {
             return Math.min(getTotalMemory(), 1536);
         } else {
